requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.1
numpy==1.24.3
scikit-learn==1.3.0
nltk==3.8.1
transformers==4.33.2
torch==2.0.1
tensorflow==2.13.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
feedparser==6.0.10
redis==4.6.0
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
boto3>=1.34.0